+++ server/component/BaseHooks.php
@@ -161,5 +161,19 @@ class BaseHooks extends BaseModel
         $method->setAccessible(false);
         return $res;
     }
+
+    /**
+     * Get the plugin version
+     */
+    public function get_plugin_db_version($plugin_name = null)
+    {
+        $res = $this->services->get_db()->query_db_first(
+            'SELECT `version` FROM `plugins` WHERE `name` = :plugin_name',
+            array(
+                ":plugin_name" => $plugin_name
+            )
+        );
+        return $res && $res['version']  ? $res['version'] : 'no data';
+    }
 }
 ?>